requirements.txt
flask==2.2.5
flask-cors==3.0.10
sqlalchemy==2.0.21
pillow==10.0.0
numpy==1.26.4
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
